Standardize AppendFormat between StringBuilder and ValueStringBuilder
