+++ src/libraries/System.Private.CoreLib/src/System/String.Manipulation.cs
@@ -540,7 +540,7 @@ namespace System
 
             var sb = new ValueStringBuilder(stackalloc char[StackallocCharBufferSizeLimit]);
             sb.EnsureCapacity(format.Length + args.Length * 8);
-            sb.AppendFormatHelper(provider, format, args);
+            sb.AppendFormat(provider, format, args);
             return sb.ToString();
         }
